paasio: refactor/simplify test file(s)
